[featute][c1]: improve

+++ src/main/java/com/example/university/model/ChuyenNganh.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ChuyenNganh {
-    private Integer maCN;
+    private Long maCN;
     private String tenCN;
     private Integer maKhoa;
 }
