+++ src/main/java/com/example/university/service/DangKyService.java
@@ -10,11 +10,11 @@ public class DangKyService {
     @Autowired
     private DangKyRepository repo;
 
-    public void dangKyHoc(int maSV, String maLHP) {
+    public void dangKyHoc(Long maSV, String maLHP) {
         repo.dangKyHoc(maSV, maLHP);
     }
 
-    public boolean huyDangKyHoc(int maSV, String maLHP) {
+    public boolean huyDangKyHoc(Long maSV, String maLHP) {
         try {
             repo.huyDangKyHoc(maSV, maLHP);
             return true;
